Extract contract extract code to a separate class
